Import LibraryEntries as posts
Each LibraryEntry has a title and a body and can have multiple files
attached.

We're going to create one post for each LibraryEntry, and then upload
each of its files against it, as another stage in the import following
this one.

Each Library is associated with one Community; we find the category to
use for the post by looking at the Discussion associated with that same
Community.