Tag imported LibraryEntries with `library`
This should make it easier for the admins to see how the import has run,
and to provide a somewhat equivalent user experience for browsing files
if they want to.

If they don't want tags it's straightforward to delete the tag, which
will remove it from any tagged post.

We'll be able to use this same approach for Announcement and Blog
imports.